Remove lock from OptimisticNormalizedCache and revert merging OptimisticNormalizedCache and MemoryCache
